Add support for encrypting vehicle responses
Clients indicate that they support encrypted vehhicle responses by
setting a flag bit in the request. The flag is authenticated but ignored
by vehicles that do not support encryption.

Clients should always set the flag when possible. In the future,
vehicles may refuse to respond to certain requests if the client does
not support this feature.